+++ tests/modeling/tpc_ds_duckdb/aggregates/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -1,4 +1,4 @@
 [query_03]
-parse_time = 0.063099
-exec_time = 0.064424
-comp_time = 0.589314
+parse_time = 0.05632
+exec_time = 0.051607
+comp_time = 0.169749
