+++ tests/modeling/tpc_ds_duckdb/tcp-ds-summary.md
@@ -25,23 +25,23 @@ Top 5 queries where PreQL is longest vs reference SQL
 | 29 | 1,526 | 1,089 | +437 |
 | 35 | 2,175 | 1,745 | +430 |
 
-Trilogy execution is faster than the reference SQL for 43/99 queries. Total Trilogy execution time is 15.314s vs 64.462s reference SQL time.
+Trilogy execution is faster than the reference SQL for 45/99 queries. Total Trilogy execution time is 12.742s vs 57.542s reference SQL time.
 
 | Performance metric | P10 | P50 | P90 |
 | --- | ---: | ---: | ---: |
-| Trilogy - Reference SQL seconds | -0.126s | +0.002s | +0.091s |
-| Trilogy vs Reference SQL | -57.8% | +4.8% | +201.7% |
-| Trilogy / Reference SQL | 0.42x | 1.05x | 3.02x |
+| Trilogy - Reference SQL seconds | -0.121s | +0.002s | +0.066s |
+| Trilogy vs Reference SQL | -59.0% | +4.0% | +195.1% |
+| Trilogy / Reference SQL | 0.41x | 1.04x | 2.95x |
 
 Top 5 queries where reference SQL is fastest vs Trilogy
 
 | Query | Trilogy s | Reference SQL s | Trilogy - Reference SQL |
 | --- | ---: | ---: | ---: |
-| 05 | 0.672s | 0.096s | +0.576s |
-| 23 | 0.722s | 0.340s | +0.382s |
-| 73 | 0.324s | 0.034s | +0.290s |
-| 76 | 0.298s | 0.057s | +0.240s |
-| 28 | 0.211s | 0.047s | +0.165s |
+| 05 | 0.615s | 0.059s | +0.556s |
+| 23 | 0.562s | 0.285s | +0.277s |
+| 73 | 0.265s | 0.031s | +0.233s |
+| 76 | 0.259s | 0.053s | +0.206s |
+| 28 | 0.208s | 0.043s | +0.165s |
 
 ## Alternative Queries
 
@@ -60,18 +60,18 @@ Top 5 queries where PreQL is longest vs reference SQL
 | --- | ---: | ---: | ---: |
 | 30.alt | 1,662 | 1,507 | +155 |
 
-Trilogy execution is faster than the reference SQL for 2/5 queries. Total Trilogy execution time is 2.990s vs 0.353s reference SQL time.
+Trilogy execution is faster than the reference SQL for 2/5 queries. Total Trilogy execution time is 2.530s vs 0.291s reference SQL time.
 
 | Performance metric | P10 | P50 | P90 |
 | --- | ---: | ---: | ---: |
-| Trilogy - Reference SQL seconds | -0.015s | +0.025s | +1.323s |
-| Trilogy vs Reference SQL | -15.5% | +49.9% | +2255.9% |
-| Trilogy / Reference SQL | 0.84x | 1.50x | 23.56x |
+| Trilogy - Reference SQL seconds | -0.018s | +0.025s | +1.125s |
+| Trilogy vs Reference SQL | -23.5% | +83.5% | +2098.9% |
+| Trilogy / Reference SQL | 0.76x | 1.83x | 21.99x |
 
 Top 5 queries where reference SQL is fastest vs Trilogy
 
 | Query | Trilogy s | Reference SQL s | Trilogy - Reference SQL |
 | --- | ---: | ---: | ---: |
-| 97.2 | 1.396s | 0.058s | +1.338s |
-| 97.1 | 1.360s | 0.059s | +1.301s |
-| 30.alt | 0.076s | 0.051s | +0.025s |
+| 97.1 | 1.181s | 0.054s | +1.127s |
+| 97.2 | 1.176s | 0.053s | +1.122s |
+| 30.alt | 0.056s | 0.030s | +0.025s |
