+++ tests/modeling/tpc_ds_duckdb/zquery41.log
@@ -1,6 +1,6 @@
 query_id = "41"
 generator = "v3"
-gen_length = 2033
+gen_length = 2079
 preql_size = 1244
 comp_size = 1166
 generated_sql = """
@@ -17,13 +17,14 @@ GROUP BY
     2),
 cheerful as (
 SELECT
-    \"item_items\".\"I_MANUFACT\" as \"item_manufact\"
+    \"item_items\".\"I_MANUFACT\" as \"item_manufact\",
+    count(CASE WHEN ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'tan' and \"item_items\".\"I_UNITS\" = 'Oz' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'purple' and \"item_items\".\"I_UNITS\" = 'Ton' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Men' and \"item_items\".\"I_COLOR\" = 'misty' and \"item_items\".\"I_UNITS\" = 'Box' and \"item_items\".\"I_SIZE\" = 'medium' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'medium' and \"item_items\".\"I_UNITS\" = 'Tsp' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'midnight' and \"item_items\".\"I_UNITS\" = 'Gram' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'pale' and \"item_items\".\"I_UNITS\" = 'Pound' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'khaki' and \"item_items\".\"I_UNITS\" = 'Pallet' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'mint' and \"item_items\".\"I_UNITS\" = 'Gross' and \"item_items\".\"I_SIZE\" = 'N/A' ) THEN \"item_items\".\"I_ITEM_SK\" ELSE NULL END) as \"manufact_matches\"
 FROM
     \"memory\".\"item\" as \"item_items\"
 GROUP BY
     1
 HAVING
-    count(CASE WHEN ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'tan' and \"item_items\".\"I_UNITS\" = 'Oz' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'purple' and \"item_items\".\"I_UNITS\" = 'Ton' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Men' and \"item_items\".\"I_COLOR\" = 'misty' and \"item_items\".\"I_UNITS\" = 'Box' and \"item_items\".\"I_SIZE\" = 'medium' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'medium' and \"item_items\".\"I_UNITS\" = 'Tsp' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'midnight' and \"item_items\".\"I_UNITS\" = 'Gram' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Books' and \"item_items\".\"I_COLOR\" = 'pale' and \"item_items\".\"I_UNITS\" = 'Pound' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'khaki' and \"item_items\".\"I_UNITS\" = 'Pallet' and \"item_items\".\"I_SIZE\" = 'N/A' ) or ( \"item_items\".\"I_CATEGORY\" = 'Electronics' and \"item_items\".\"I_COLOR\" = 'mint' and \"item_items\".\"I_UNITS\" = 'Gross' and \"item_items\".\"I_SIZE\" = 'N/A' ) THEN \"item_items\".\"I_ITEM_SK\" ELSE NULL END) > 0
+    \"manufact_matches\" > 0
 )
 SELECT
     \"highfalutin\".\"filtered_product_name\" as \"filtered_product_name\"
