+++ tests/modeling/tpc_h/tcp-h-summary.md
@@ -19,20 +19,20 @@ Top 5 queries where PreQL is longest vs reference SQL
 
 None.
 
-Trilogy execution is faster than the reference SQL for 6/22 queries. Total Trilogy execution time is 0.866s vs 0.659s reference SQL time.
+Trilogy execution is faster than the reference SQL for 8/22 queries. Total Trilogy execution time is 0.653s vs 0.496s reference SQL time.
 
 | Performance metric | P10 | P50 | P90 |
 | --- | ---: | ---: | ---: |
-| Trilogy - Reference SQL seconds | -0.013s | +0.005s | +0.032s |
-| Trilogy vs Reference SQL | -28.8% | +29.8% | +98.3% |
-| Trilogy / Reference SQL | 0.71x | 1.30x | 1.98x |
+| Trilogy - Reference SQL seconds | -0.010s | +0.004s | +0.016s |
+| Trilogy vs Reference SQL | -29.2% | +26.0% | +113.8% |
+| Trilogy / Reference SQL | 0.71x | 1.26x | 2.14x |
 
 Top 5 queries where reference SQL is fastest vs Trilogy
 
 | Query | Trilogy s | Reference SQL s | Trilogy - Reference SQL |
 | --- | ---: | ---: | ---: |
-| 20 | 0.093s | 0.024s | +0.069s |
-| 17 | 0.060s | 0.013s | +0.047s |
-| 16 | 0.082s | 0.049s | +0.033s |
-| 21 | 0.084s | 0.066s | +0.018s |
-| 11 | 0.050s | 0.033s | +0.017s |
+| 20 | 0.075s | 0.015s | +0.060s |
+| 17 | 0.056s | 0.009s | +0.047s |
+| 16 | 0.043s | 0.027s | +0.016s |
+| 04 | 0.026s | 0.012s | +0.014s |
+| 13 | 0.039s | 0.028s | +0.011s |
