+++ tests/modeling/tpc_h/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -1,109 +1,109 @@
 [query_01]
-parse_time = 0.019299099999898317
-exec_time = 0.042212699999936376
-comp_time = 0.044156100000009246
+parse_time = 0.018008699999882083
+exec_time = 0.038748499999655905
+comp_time = 0.03881229999979041
 
 [query_02]
-parse_time = 0.03267090000008466
-exec_time = 0.02690529999995306
-comp_time = 0.01989800000001196
+parse_time = 0.03327839999974458
+exec_time = 0.016580599999997503
+comp_time = 0.012017300000024989
 
 [query_03]
-parse_time = 0.06058189999998831
-exec_time = 0.031318499999997584
-comp_time = 0.030812400000058915
+parse_time = 0.05831900000021051
+exec_time = 0.024710299999696872
+comp_time = 0.025211600000147882
 
 [query_04]
-parse_time = 0.026652699999885954
-exec_time = 0.031534099999817045
-comp_time = 0.01562950000015917
+parse_time = 0.02554099999997561
+exec_time = 0.025567799999862473
+comp_time = 0.011799899999914487
 
 [query_05]
-parse_time = 0.04489630000011857
-exec_time = 0.05160650000016176
-comp_time = 0.03553280000005543
+parse_time = 0.044593499999791675
+exec_time = 0.04046199999993405
+comp_time = 0.02988920000007056
 
 [query_06]
-parse_time = 0.01720049999994444
-exec_time = 0.00377529999991566
-comp_time = 0.0034051000000090426
+parse_time = 0.017291900000145688
+exec_time = 0.0035144000003128895
+comp_time = 0.003240400000322552
 
 [query_07]
-parse_time = 0.04968060000010155
-exec_time = 0.02508920000013859
-comp_time = 0.024682100000063656
+parse_time = 0.04900399999996807
+exec_time = 0.01932780000015555
+comp_time = 0.01986129999977493
 
 [query_08]
-parse_time = 0.05512100000009923
-exec_time = 0.04220190000000912
-comp_time = 0.04324339999993754
+parse_time = 0.05348890000004758
+exec_time = 0.03526689999989685
+comp_time = 0.03699710000000778
 
 [query_09]
-parse_time = 0.04057420000003731
-exec_time = 0.032218400000147085
-comp_time = 0.047311300000046685
+parse_time = 0.03757850000010876
+exec_time = 0.025890699999763456
+comp_time = 0.03847799999994095
 
 [query_10]
-parse_time = 0.04352360000007138
-exec_time = 0.03000459999998384
-comp_time = 0.04383589999997639
+parse_time = 0.042702100000042265
+exec_time = 0.02295639999965715
+comp_time = 0.03372590000026321
 
 [query_11]
-parse_time = 0.023811600000044564
-exec_time = 0.05026459999999133
-comp_time = 0.03316580000000613
+parse_time = 0.023726799999622017
+exec_time = 0.02539230000002135
+comp_time = 0.016711100000065926
 
 [query_12]
-parse_time = 0.02476709999996274
-exec_time = 0.023473000000194588
-comp_time = 0.023993900000050417
+parse_time = 0.025287899999966612
+exec_time = 0.019793599999957223
+comp_time = 0.019841399999677378
 
 [query_13]
-parse_time = 0.01875489999997626
-exec_time = 0.04782690000001821
-comp_time = 0.03623270000002776
+parse_time = 0.01813369999990755
+exec_time = 0.03869880000002013
+comp_time = 0.028038300000389427
 
 [query_14]
-parse_time = 0.037963300000001254
-exec_time = 0.009045100000093953
-comp_time = 0.005998900000122376
+parse_time = 0.0380645000000186
+exec_time = 0.0072859000001699314
+comp_time = 0.005532099999982165
 
 [query_15]
-parse_time = 0.03065839999999298
-exec_time = 0.02035559999990255
-comp_time = 0.013366700000005949
+parse_time = 0.030803999999989173
+exec_time = 0.01538250000021435
+comp_time = 0.008195599999908154
 
 [query_16]
-parse_time = 0.021513400000003458
-exec_time = 0.08230179999986831
-comp_time = 0.04923660000008567
+parse_time = 0.021763999999620864
+exec_time = 0.043206499999996595
+comp_time = 0.027475099999719532
 
 [query_17]
-parse_time = 0.039805999999998676
-exec_time = 0.060067500000059226
-comp_time = 0.012845900000002075
+parse_time = 0.03908799999999246
+exec_time = 0.05606830000033369
+comp_time = 0.009367099999963102
 
 [query_18]
-parse_time = 0.03129609999996319
-exec_time = 0.030236600000080216
-comp_time = 0.0490273000000343
+parse_time = 0.03061279999974431
+exec_time = 0.02536069999996471
+comp_time = 0.040348699999867677
 
 [query_19]
-parse_time = 0.04477109999993445
-exec_time = 0.025342199999840886
-comp_time = 0.02163369999993847
+parse_time = 0.042204599999877246
+exec_time = 0.023333000000093307
+comp_time = 0.020878199999970093
 
 [query_20]
-parse_time = 0.0332544999998845
-exec_time = 0.09343379999995705
-comp_time = 0.024028600000065126
+parse_time = 0.0328337000000829
+exec_time = 0.07493849999991653
+comp_time = 0.014521200000217505
 
 [query_21]
-parse_time = 0.03788319999989653
-exec_time = 0.08415500000000975
-comp_time = 0.06599479999999858
+parse_time = 0.03637939999998707
+exec_time = 0.05734759999995731
+comp_time = 0.04767959999981031
 
 [query_22]
-parse_time = 0.03208429999995133
-exec_time = 0.022330600000032064
-comp_time = 0.014977599999838276
+parse_time = 0.031550700000025245
+exec_time = 0.013406799999756913
+comp_time = 0.007566700000097626
